Rgb stores values as doubles
Rgb now stores the r, g, b components as doubles to avoid rounding
issues that could occur when converting between color spaces.

Rgb now clamps r, g, b values to 0.0 - 250.0 when converting to a JavaFX
color. This matches CSS rgb() behavior where negative values are treated
as 0.

+++ kubed-core/src/main/kotlin/kubed/color/Hcg.kt
@@ -68,6 +68,6 @@ class Hcg(val h: Double, var c: Double, var g: Double, var opacity: Double = 1.0
     override fun darker(k: Double) = rgb().darker().hcg()
 
     private fun hcg2rgb(r: Double, g: Double, b: Double, m: Double, a: Double): Rgb {
-        return Rgb(((r + m) * 255).toInt(), ((g + m) * 255).toInt(), ((b + m) * 255).toInt(), a)
+        return Rgb((r + m) * 255, (g + m) * 255, (b + m) * 255, a)
     }
 }
